requirements.txt
discord.py
pytz
selenium